Add content, still very unstructured, and some preliminary styling
Remove most of the styles from the impress demo, we won't need them and
it makes the css file a bit easier to read. But use the nice .slide
frame from the demo for all slides.

A general style for the slides with little content would be nice, so
that the text is large and in the middle.

And most important would be to get it more structured and the content
straight ...

And maybe some photos / illustrations?